GetFailedOperationsIdmResourceTest test is extracted
